Added Downloading to recycler view , not working tho
