+++ app/src/main/java/sid/comslav/com/circleofmusic/TrackListAdapter.java
@@ -1,11 +1,14 @@
 package sid.comslav.com.circleofmusic;
 
+import android.app.DownloadManager;
+import android.content.Context;
 import android.graphics.Color;
+import android.net.Uri;
+import android.os.Environment;
 import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.ImageView;
 import android.widget.TextView;
 
 public class TrackListAdapter extends RecyclerView.Adapter<TrackListAdapter.ViewHolder> {
@@ -31,7 +34,7 @@ public class TrackListAdapter extends RecyclerView.Adapter<TrackListAdapter.View
         if (newNotification[position]) {
             holder.mTextView.setTextColor(Color.GREEN);
         }
-        holder.mImageView.setImageResource(R.mipmap.ic_launcher);
+        holder.sTextView.setText("Track Info After Download");
     }
 
     @Override
@@ -39,14 +42,29 @@ public class TrackListAdapter extends RecyclerView.Adapter<TrackListAdapter.View
         return mDataSet.length;
     }
 
-    public static class ViewHolder extends RecyclerView.ViewHolder {
+    public static class ViewHolder extends RecyclerView.ViewHolder implements View.OnClickListener {
         public TextView mTextView;
-        public ImageView mImageView;
+        public TextView sTextView;
 
         public ViewHolder(View view) {
             super(view);
             this.mTextView = (TextView) view.findViewById(R.id.tvTrackName);
-            this.mImageView = (ImageView) view.findViewById(R.id.ivTrackArt);
+            this.sTextView = (TextView) view.findViewById(R.id.tvTrackInfo);
+        }
+
+        @Override
+        public void onClick(View v) {
+            downloadMusicTrack(v.toString());
+        }
+
+        void downloadMusicTrack(String selectedItem) {
+            String url = "http://circleofmusic-sidzi.rhcloud.com/downloadTrack" + selectedItem;
+            DownloadManager.Request request = new DownloadManager.Request(Uri.parse(url));
+            request.setDescription("Downloading");
+            request.setTitle(selectedItem);
+            request.setDestinationInExternalPublicDir(Environment.DIRECTORY_DOWNLOADS, selectedItem);
+            DownloadManager manager = (DownloadManager) getSystemService(Context.DOWNLOAD_SERVICE);
+            manager.enqueue(request);
         }
     }
 }
